refactor(util): move GitHub token related utils to proper place
originated from: https://github.com/renovatebot/renovate/pull/23810#discussion_r1290946219
